Remove alexa-conversation module and creater helper module
